+++ src/api/status/tests/test_index.py
@@ -1,5 +1,8 @@
+import json
 import os
 from unittest import mock
+
+from nhs_context_logging import app_logger
 
 
 def test_index():
@@ -7,8 +10,40 @@ def test_index():
         from api.status.index import handler
 
         result = handler(event={})
+
+    expected_body = json.dumps(
+        {
+            "resourceType": "OperationOutcome",
+            "id": app_logger.service_name,
+            "meta": {
+                "profile": [
+                    "https://fhir.nhs.uk/StructureDefinition/NHSDigital-OperationOutcome"
+                ]
+            },
+            "issue": [
+                {
+                    "severity": "information",
+                    "code": "informational",
+                    "details": {
+                        "coding": [
+                            {
+                                "system": "https://fhir.nhs.uk/StructureDefinition/NHSDigital-OperationOutcome",
+                                "code": "OK",
+                                "display": "Transaction successful",
+                            }
+                        ]
+                    },
+                    "diagnostics": "Transaction successful",
+                }
+            ],
+        }
+    )
+
     assert result == {
         "statusCode": 200,
-        "body": "OK",
-        "headers": {"Content-Length": 2, "Content-Type": "application/json"},
+        "body": expected_body,
+        "headers": {
+            "Content-Length": len(expected_body),
+            "Content-Type": "application/json",
+        },
     }